Work around snap confinement for LD_LIBRARY_PATH
LD_LIBRARY_PATH gets appended with all of the core24 and checkbox24
paths, which is complicating our testing landscape and causing warnings
in tests. This is because those versions only match if running on Noble,
but we validate the graphics stack on newer releases and want to test
the system itself.

Specifically, libxml2 versions between Noble and Oracular were causing
Crucible to throw warnings